Adjust panel hiding algorithm to account leave-event not firing
On GNOME 42, leave-event doesn't fire if the mouse cursor doesn't touch
the desktop immediately after leaving the panel. Adjust the hiding
algorithm to account for this by using PointerWatcher to monitor cursor
movements. Check if the cursor is still on the panel manually using the
reported coordinates and hide the panel when it leaves.

Fixes #354.